Extract common storage file system clients code
